requirements.txt
# Web Framework
flask==2.0.1
flask-cors==3.0.10
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
gunicorn==20.1.0

# Database
pymongo==4.0.1

# API and HTTP
requests==2.25.1
betfairlightweight==2.12.0
requests-oauthlib==1.3.0

# Authentication & Security
pyjwt==2.3.0
bcrypt==3.2.0

# Environment & Configuration
python-dotenv==0.19.1

# Utilities
python-dateutil==2.8.2
six==1.16.0

# Documentation
flask-swagger-ui==3.36.0
